make env generated automatically

+++ generate_env.py
@@ -1,8 +1,9 @@
 import secrets
 import os
 
-
-secret_key = secrets.token_hex(32)
-
-with open('.env', 'w') as f:
-    f.write(f'SECRET_KEY={secret_key}\n')
+def generate_env():
+    if not os.path.exists('.env'):
+        secret_key = secrets.token_hex(32)
+        with open('.env', 'w') as f:
+            f.write(f'SECRET_KEY={secret_key}\n')
+            f.write(f"DATABASE_URL = sqlite:///unimap.db")
